LayoutTests/http/tests/media/video-play-suspend-expected.txt

Test that timeupdate events are sent when media loading suspends itself.

RUN(video.play())
EVENT(loadstart)
EVENT(suspend)
EVENT(timeupdate)
END OF TEST
